+++ lib/xcodeproj/project/object.rb
@@ -65,7 +65,9 @@ module Xcodeproj
           @project, @uuid = project, uuid
           @isa = self.class.isa
           @referrers = []
-          raise "[Xcodeproj] Attempt to initialize an abstract class." unless @isa.match(/^(PBX|XC)/)
+          unless @isa.match(/^(PBX|XC)/)
+            raise "[Xcodeproj] Attempt to initialize an abstract class."
+          end
         end
 
         # Initializes the object with the default values of simple attributes.
@@ -261,7 +263,10 @@ module Xcodeproj
         def configure_with_plist(objects_by_uuid_plist)
           object_plist = objects_by_uuid_plist[uuid].dup
 
-          raise "[Xcodeproj] Attempt to initialize `#{isa}` from plist with different isa `#{object_plist}`" unless object_plist['isa'] == isa
+          unless object_plist['isa'] == isa
+          raise "[Xcodeproj] Attempt to initialize `#{isa}` from plist with " \
+            "different isa `#{object_plist}`"
+          end
           object_plist.delete('isa')
 
           simple_attributes.each do |attrb|
@@ -303,9 +308,10 @@ module Xcodeproj
           end
 
           unless object_plist.empty?
-            raise "[!] Xcodeproj doesn't know about the following attributes " \
-                  "#{object_plist.inspect} for the '#{isa}' isa.\n" \
-                  "Please file an issue: https://github.com/CocoaPods/Xcodeproj/issues/new"
+            raise "[!] Xcodeproj doesn't know about the following " \
+                  "attributes #{object_plist.inspect} for the '#{isa}' isa." \
+                  "\nPlease file an issue: " \
+                  "https://github.com/CocoaPods/Xcodeproj/issues/new"
           end
         end
 
@@ -333,15 +339,15 @@ module Xcodeproj
         def object_with_uuid(uuid, objects_by_uuid_plist, attribute)
           unless object = project.objects_by_uuid[uuid] || project.new_from_plist(uuid, objects_by_uuid_plist)
             UI.warn "`#{inspect}` attempted to initialize an object with " \
-              "an unknown UUID. `#{uuid}` for attribute: `#{attribute.name}`."\
-              " This can be the result of a merge and the unknown UUID is "    \
-              "being discarded."
+              "an unknown UUID. `#{uuid}` for attribute: " \
+              "`#{attribute.name}`. This can be the result of a merge and  " \
+              "the unknown UUID is being discarded."
           end
           object
         rescue NameError
           attributes = objects_by_uuid_plist[uuid]
           raise "`#{isa}` attempted to initialize an object with unknown ISA "\
-                "`#{attributes['isa']}` from attributes: `#{attributes}`\n"   \
+                "`#{attributes['isa']}` from attributes: `#{attributes}`\n" \
                 "Please file an issue: https://github.com/CocoaPods/Xcodeproj/issues/new"
         end
 
@@ -460,9 +466,10 @@ module Xcodeproj
   end
 end
 
-require 'xcodeproj/project/object_list'
-require 'xcodeproj/project/object_dictionary'
+require 'xcodeproj/project/name_helper'
 require 'xcodeproj/project/object_attributes'
+require 'xcodeproj/project/object_dictionary'
+require 'xcodeproj/project/object_list'
 
 # Required because some classes have cyclical references to each other.
 #