+++ spec/project/object_attributes_spec.rb
@@ -23,37 +23,20 @@ module ProjectSpecs
       @attribute.plist_name.should == 'sourceTree'
     end
 
-    it "caches the plist names" do
-      AbstractObjectAttribute.stubs(:plist_name_store).returns({})
-      String.any_instance.expects(:camelize).returns('sourceTree')
-      @attribute.plist_name
-      String.any_instance.expects(:camelize).never
-      @attribute.plist_name.should == 'sourceTree'
-    end
 
-    it "can store the accepted classes for the value type checking" do
-      @attribute.classes = [String]
-      @attribute.classes.should == [String]
-    end
-
-    it "can store a default value" do
-      @attribute.default_value = 'A_ROOT'
-      @attribute.default_value.should == 'A_ROOT'
-    end
-
-    it "can get its value for a given object" do
+    it "returns its value for a given object" do
       file = @project.new(PBXFileReference)
       file.source_tree = 'A_ROOT'
       @attribute.get_value(file).should == 'A_ROOT'
     end
 
-    it "can set its value for a given object" do
+    it "sets its value for a given object" do
       file = @project.new(PBXFileReference)
       @attribute.set_value(file, 'A_ROOT')
       file.source_tree.should == 'A_ROOT'
     end
 
-    it "can set its default value for a given object" do
+    it "sets its default value for a given object" do
       @attribute.default_value = 'SAMPLE_ROOT'
       file = @project.new(PBXFileReference)
       file.source_tree.should.not == 'SAMPLE_ROOT'
@@ -61,19 +44,52 @@ module ProjectSpecs
       file.source_tree.should == 'SAMPLE_ROOT'
     end
 
-    it "can validate a simple value" do
+    it "validates a simple value" do
       @attribute.classes = [String]
       lambda { @attribute.validate_value('a string') }.should.not.raise
       lambda { @attribute.validate_value(['array']) }.should.raise
     end
 
-    it "can validate an xcodeproj object isa" do
+    it "validates an xcodeproj object ISA" do
       attrb = AbstractObjectAttribute.new(:to_many, :children, PBXGroup)
       attrb.classes = [PBXFileReference, PBXGroup]
       file = @project.new(PBXFileReference)
       target = @project.new(PBXNativeTarget)
       lambda { attrb.validate_value(file) }.should.not.raise
       lambda { attrb.validate_value(target) }.should.raise
+    end
+
+
+    describe "references by keys attributes" do
+      before do
+        @attribute = AbstractObjectAttribute.new(:references_by_keys, :project_references, PBXProject)
+        @attribute.classes = [PBXFileReference, PBXGroup]
+        @attribute.classes_by_key = {
+          :project_ref   => PBXFileReference,
+          :product_group => PBXGroup
+        }
+      end
+
+      it "validates the key of an attribute which stores" do
+        file = @project.new(PBXFileReference)
+        should.raise do
+          @attribute.validate_value_for_key(file, :not_allowed)
+        end.message.should.include?('unsupported key')
+      end
+
+      it "validates the ISA of the value" do
+        file = @project.new(PBXFileReference)
+        should.raise do
+          @attribute.validate_value_for_key(file, :product_group)
+        end.message.should.include?('Type checking error')
+      end
+
+      it "accepts a value" do
+        file = @project.new(PBXFileReference)
+        should.not.raise do
+          @attribute.validate_value_for_key(file, :project_ref)
+        end
+      end
     end
   end
 end