+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Master
 
-* [PlistHelper] Now the `plutil` tool is used to save the files if available to
+* `PlistHelper`: Now the `plutil` tool is used to save the files if available to
   produce output consistent with Xcode.  
   [Fabio Pelosin](https://github.com/irrationalfab)
+
+* `PBXGroup`: Now file references to Xcode projects are properly handled and
+  setup. Also the `ObjectDictionary` class has been improved and now can be
+  used to edit the attributes using it.
+  [Fabio Pelosin](https://github.com/irrationalfab)
+  [Xcodeproj#172](https://github.com/CocoaPods/Xcodeproj/pull/172)
 
 ## 0.18.0